backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Pillow==10.0.1
cryptography==41.0.3
